+++ migrations/000001_create_table_menu.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS menu (
     id CHAR(36) PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT NOT NULL,
-    price INT NOT NULL,
+    price FLOAT NOT NULL,
     category ENUM(
             'main',
             'appetizer',
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS menu (
             'spicy'
             ) NOT NULL,
     image_url VARCHAR(255) NOT NULL,
-    rating INT NOT NULL DEFAULT 0,
+    rating FLOAT NOT NULL DEFAULT 0,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     deleted BOOLEAN DEFAULT FALSE,
