+++ migrations/000003_create_table_order.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS orders (
     id CHAR(36) PRIMARY KEY,
     user_id CHAR(36) NOT NULL,
     status ENUM("pending","processing","success","failed") NOT NULL DEFAULT 'pending',
-    amount INT NOT NULL,
+    amount FLOAT NOT NULL,
     payment_status ENUM("paid", "unpaid") NOT NULL DEFAULT 'unpaid',
     payment_method VARCHAR(50),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
